@@ -15,16 +15,16 @@ classifiers =
     Programming Language :: Python :: 3.9
     Topic :: Communications :: Chat :: Internet Relay Chat
 [options]
-python_requires = >=3.6
+python_requires = >=3.7
 [options.entry_points]
 sopel.plugins =
     miraheze    = MirahezeBots.plugins.miraheze
     phab        = MirahezeBots.plugins.phab
     responses   = MirahezeBots.plugins.responses
-    rss         = MirahezeBots.plugins.rss
     shortlinks  = MirahezeBots.plugins.shortlinks
     status      = MirahezeBots.plugins.status
     welcome     = MirahezeBots.plugins.welcome
     goofy       = MirahezeBots.plugins.goofy
+    shared      = MirahezeBots.plugins.shared
 console_scripts =
-    sopel-dbclean = MirahezeBots.dbclean:main
+    sopel-dbclean = MirahezeBots.dbclean:rundel
