@@ -1,20 +1,24 @@
-"""responses.py - like a FAQ bot"""
+"""responses.py - like a FAQ bot."""
+
+from MirahezeBots.version import SHORTVERSION, VERSION
 
 from sopel.config.types import StaticSection, ValidatedAttribute
 from sopel.module import commands, example, rate, require_account
 
-from MirahezeBots.version import VERSION, SHORTVERSION
-
 
 class ResponsesSection(StaticSection):
+    """Create configuration for Sopel."""
+
     support_channel = ValidatedAttribute('support_channel', str)
 
 
 def setup(bot):
+    """Set up the config section."""
     bot.config.define_section('responses', ResponsesSection)
 
 
 def configure(config):
+    """Set up the configuration options."""
     config.define_section('responses', ResponsesSection, validate=False)
     config.responses.configure_setting('support_channel', 'Specify a support IRC channel (leave blank for none).')
 
@@ -27,10 +31,10 @@ def addchan(bot, trigger):
     """Reply to channel request message."""
     admins = ' '.join(map(str, bot.config.core.admin_accounts))
     if bot.config.responses.support_channel is not None:
-        bot.say(("Hey {}, {} would like to have me in their channel: {}").format(admins, trigger.nick, trigger.group(2)),
+        bot.say(f'Hey {admins}, {trigger.nick} would like to have me in their channel: {trigger.group(2)}',
                 bot.config.responses.support_channel)
         if trigger.sender != bot.config.responses.support_channel:
-            bot.reply("Request sent! Action upon the request should be taken shortly. Thank you for using {}!".format(bot.nick))
+            bot.reply(f'Request sent! Action upon the request should be taken shortly. Thank you for using {bot.nick}!')
 
 
 @commands('gj', 'gw')
@@ -38,7 +42,7 @@ def addchan(bot, trigger):
 @rate(user=2, channel=1, server=0)
 def gj(bot, trigger):
     """Tell the user that they are doing good work."""
-    bot.say(("You're doing good work, {}!").format(trigger.group(2)))
+    bot.say(f"You're doing good work, {trigger.group(2)}!")
 
 
 @commands('cancelreminder')
@@ -47,21 +51,20 @@ def gj(bot, trigger):
 def cancel(bot, trigger):
     """Cancel reminder."""
     admins = ' '.join(map(str, bot.config.core.admin_accounts))
-    bot.reply(('Pinging {} to cancel '
-               '{}\'s reminder.').format(admins, trigger.nick))
+    bot.reply(f"Pinging {admins} to cancel {trigger.nicks}'s reminder.")
 
 
 @commands('botversion', 'bv')
 @example('.botversion')
 @rate(user=2, channel=1, server=0)
-def botversion(bot, trigger):
+def botversion(bot, trigger):  # noqa: U100
     """List the current version of the bot."""
-    bot.say('The current version of this bot is {} ({})'.format(VERSION, SHORTVERSION))
+    bot.reply(f'The current version of this bot is {VERSION} ({SHORTVERSION})')
 
 
 @commands('source', 'botsource')
 @example('.source')
 @rate(user=2, channel=1, server=0)
-def githubsource(bot, trigger):
+def githubsource(bot, trigger):  # noqa: U100
     """Give the link to MirahezeBot's Github."""
     bot.reply('My code can be found here: https://github.com/MirahezeBots/MirahezeBots')
