@@ -1,9 +1,12 @@
-""" The functions in this file are not suitable for non-internal use. They are subject to change without notice and are not yet released. """
+"""MediaWiki API Handler."""
 import requests
 
 
-def login(url, session, username='Example', password='password'):
-    CONNECTERRMSG = "Unable to conect to wiki"
+CONNECTERRMSG = 'Unable to conect to wiki'
+
+
+def login(url, session, username, password):
+    """Login to MediaWiki API using bot password system."""
     PARAMS_0 = {
         'action': 'query',
         'meta': 'tokens',
@@ -14,7 +17,7 @@ def login(url, session, username='Example', password='password'):
         request = session.get(url=url, params=PARAMS_0)
         DATA = request.json()
     except Exception:
-        return ["Error", CONNECTERRMSG]
+        return ['Error', CONNECTERRMSG]
 
     LOGIN_TOKEN = DATA['query']['tokens']['logintoken']
 
@@ -28,24 +31,25 @@ def login(url, session, username='Example', password='password'):
     try:
         session.post(url, data=PARAMS_1)
     except Exception:
-        return ["Error", CONNECTERRMSG]
-    return ["Success", "Logged in"]
+        return ['Error', CONNECTERRMSG]
+    return ['Success', 'Logged in']
 
 
-def gettoken(url, session, type='csrftoken'):
+def gettoken(url, session, tokentype='csrftoken'):
+    """Get a token from the meta::tokens api."""
     PARAMS_2 = {'action': 'query', 'meta': 'tokens', 'format': 'json'}
 
     try:
         request = session.get(url=url, params=PARAMS_2)
         DATA = request.json()
     except Exception:
-        return ["Error", "Unable to conect to wiki"]
+        return ['Error', CONNECTERRMSG]
 
-    TOKEN = DATA['query']['tokens'][type]
-    return TOKEN
+    return DATA['query']['tokens'][tokentype]
 
 
 def makeaction(requestinfo, action, target, performer, reason, content=''):
+    """Perform an action via the ACTIONS API."""
     if action == 'edit':
         PARAMS = {
             'action': 'edit',
@@ -56,7 +60,7 @@ def makeaction(requestinfo, action, target, performer, reason, content=''):
             'bot': 'true',
             'format': 'json',
         }
-    elif action == "create":
+    elif action == 'create':
         PARAMS = {
             'action': 'edit',
             'title': target,
@@ -102,26 +106,24 @@ def makeaction(requestinfo, action, target, performer, reason, content=''):
     try:
         request = requestinfo[1].post(requestinfo[0], data=PARAMS)
         DATA = request.json()
-        if DATA.get("error") is not None:
-            return ["MWError", (DATA.get("error").get("info"))]
-        else:
-            return ["Success", ("{0} request sent. You may want to check the {0} log to be sure that it worked.").format(action)]
+        if DATA.get('error') is not None:
+            return ['MWError', (DATA.get('error').get('info'))]
+        return ['Success', f'{action} request sent. You may want to check the {action} log to be sure that it worked.']
     except Exception:
-        return ["Fatal", ("An unexpected error occurred. Did you type the wiki or user incorrectly? Do I have {} rights on that wiki?").format(action)]
+        return [
+            'Fatal',
+            f'An unexpected error occurred. Did you type the wiki or user incorrectly? Do I have {action} on that wiki?',
+            ]  # noqa: JS102
 
 
-def main(performer, target, action, reason, url, authinfo, content=False):
-    session = requests.Session()
+def main(performer, target, action, reason, url, authinfo, content=False, session=requests.Session()):
+    """Execute a full API Sequence."""
     lg = login(url, session, authinfo[0], authinfo[1])
-    if lg[0] == "Error":
+    if lg[0] == 'Error':
         return lg[1]
-    else:
-        TOKEN = gettoken(url, session, type='csrftoken')
-        if TOKEN[0] == "Error":
-            return TOKEN[1]
-        else:
-            if content:
-                act = makeaction([url, session, TOKEN], action, target, performer, reason, content)
-            else:
-                act = makeaction([url, session, TOKEN], action, target, performer, reason)
-            return act[1]
+    TOKEN = gettoken(url, session, tokentype='csrftoken')
+    if TOKEN[0] == 'Error':
+        return TOKEN[1]
+    if content:
+        return makeaction([url, session, TOKEN], action, target, performer, reason, content)[1]
+    return makeaction([url, session, TOKEN], action, target, performer, reason)[1]
