@@ -1,13 +1,13 @@
-"""This plugin expands links to various websites"""
+"""This plugin expands links to various websites."""
 from sopel.module import commands, example
 
 
 @commands('github', 'gh')
 @example('.github user')
 def ghuser(bot, trigger):
-    """Expands a link to github."""
+    """Expand a link to github."""
     try:
-        bot.say("https://github.com/" + trigger.group(2))
+        bot.say(f'https://github.com/{trigger.group(2)}')
     except TypeError:
         bot.say('Syntax: .github user', trigger.sender)
 
@@ -15,9 +15,9 @@ def ghuser(bot, trigger):
 @commands('redditu')
 @example('.redditu example')
 def redditu(bot, trigger):
-    """Expands a link to reddit/u."""
+    """Expand a link to reddit/u."""
     try:
-        bot.say("https://reddit.com/u/" + trigger.group(2))
+        bot.say(f'https://reddit.com/u/{trigger.group(2)}')
     except TypeError:
         bot.say('Syntax: .redditu example', trigger.sender)
 
@@ -25,9 +25,9 @@ def redditu(bot, trigger):
 @commands('subred')
 @example('.subred example')
 def redditr(bot, trigger):
-    """Expands a link to reddit/r."""
+    """Expand a link to reddit/r."""
     try:
-        bot.say("https://reddit.com/r/" + trigger.group(2))
+        bot.say(f'https://reddit.com/r/{trigger.group(2)}')
     except TypeError:
         bot.say('Syntax: .subred example', trigger.sender)
 
@@ -35,10 +35,10 @@ def redditr(bot, trigger):
 @commands('wmca')
 @example('.wmca example')
 def wmca(bot, trigger):
-    """Expands a link to Wikimedia CentralAuth."""
+    """Expand a link to Wikimedia CentralAuth."""
     try:
-        target = trigger.group(2).replace(" ", "_")
-        bot.say("https://meta.wikimedia.org/wiki/Special:CentralAuth/" + target)
+        target = trigger.group(2).replace(' ', '_')
+        bot.say(f'https://meta.wikimedia.org/wiki/Special:CentralAuth/{target}')
     except AttributeError:
         bot.say('Syntax: .wmca example', trigger.sender)
 
@@ -46,10 +46,10 @@ def wmca(bot, trigger):
 @commands('mhca')
 @example('.mhca example')
 def mhca(bot, trigger):
-    """Expands a link to Miraheze Central Auth."""
+    """Expand a link to Miraheze Central Auth."""
     try:
-        target = trigger.group(2).replace(" ", "_")
-        bot.say("https://meta.miraheze.org/wiki/Special:CentralAuth/" + target)
+        target = trigger.group(2).replace(' ', '_')
+        bot.say(f'https://meta.miraheze.org/wiki/Special:CentralAuth/{target}')
     except AttributeError:
         bot.say('Syntax: .mhca example', trigger.sender)
 
@@ -57,9 +57,9 @@ def mhca(bot, trigger):
 @commands('tw')
 @example('.tw user')
 def twlink(bot, trigger):
-    """Expands a link to Twitter."""
+    """Expand a link to Twitter."""
     try:
-        bot.say("https://twitter.com/" + trigger.group(2))
+        bot.say(f'https://twitter.com/{trigger.group(2)}')
     except TypeError:
         bot.say('Syntax: .tw user', trigger.sender)
 
@@ -67,17 +67,17 @@ def twlink(bot, trigger):
 @commands('mh')
 @example('.mh wiki page')
 def mhwiki(bot, trigger):
-    """Expands a link to Miraheze wikis."""
+    """Expand a link to Miraheze wikis."""
     try:
-        options = trigger.group(2).split(" ", 1)
+        options = trigger.group(2).split(' ', 1)
         if len(options) == 1:
             page = options[0]
-            page = page.replace(" ", "_")
-            bot.say("https://meta.miraheze.org/wiki/" + page)
+            page = page.replace(' ', '_')
+            bot.say(f'https://meta.miraheze.org/wiki/{page}')
         elif len(options) == 2:
             wiki = options[0]
             page = options[1]
-            page = page.replace(" ", "_")
-            bot.say("https://" + wiki + ".miraheze.org/wiki/" + page)
+            page = page.replace(' ', '_')
+            bot.say(f'https://{wiki}.miraheze.org/wiki/{page}')
     except AttributeError:
         bot.say('Syntax: .mh wiki page', trigger.sender)
