@@ -1,3 +1,4 @@
+"""Provides version information."""
 VERSION = '9.1.3'
-VERSIONARRAY = VERSION.split(".")
+VERSIONARRAY = VERSION.split('.')
 SHORTVERSION = 'v' + str(VERSIONARRAY[0])
